Moves service to group module
