+++ src/app/modules/group/http-services/remove-group-manager.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
-import { appConfig } from '../../shared/helpers/config';
-import { SimpleActionResponse } from '../../shared/http-services/action-response';
+import { appConfig } from '../../../shared/helpers/config';
+import { SimpleActionResponse } from '../../../shared/http-services/action-response';
 
 @Injectable({
   providedIn: 'root',
